% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRS.R
\name{BRS}
\alias{BRS}
\alias{BRS_impute}
\title{Main BRS subclass predictor}
\usage{
BRS(
  x,
  train = BRSpred::CohortA.vst,
  pamrobj = BRSpred::pamr3cl,
  threshold = 0.6606584,
  type = "class",
  qnormalize = TRUE,
  zscale = TRUE,
  verb = TRUE,
  ...
)

BRS_impute(x, train, verb = TRUE)
}
\arguments{
\item{x}{New data matrix 'x' with potentially missing rownames}

\item{train}{Original training data matrix}

\item{pamrobj}{pamr-package object of class 'pamr', by default the original one trained for the package}

\item{threshold}{pamr threshold parameter (by default the optimal threshold found by CV in training data)}

\item{type}{Type of prediction as given to pamr; eligible values 'class', 'posterior', 'centroid', 'nonzero'}

\item{qnormalize}{Should quantile normalization be applied to 'x' in respect to the training data}

\item{zscale}{Should z-score scaling be applied to data; by default TRUE}

\item{verb}{Verbosity}

\item{...}{Additional parameters passed on to pamr::pamr.predict}
}
\value{
pamr::pamr.predict-call predictions for the input 'x'
}
\description{
Main BRS subclass predictor

Imputation based on medians from training data for gene names that are missing from input data
}
\details{
This is the main BCG response subtype predictor function. It essentially works as a wrapper for the original 3-class pamr-object, with imputation and other convenience functionality.

Imputation of gene expression values based on median from an existing training data.
}
\examples{
library(BRSpred)
BRS(x=CohortB.vst, train=CohortA.vst, threshold=0.6606584)

}
\references{
de Jong F. C., Laajala T. D., et al. Citation
}
