% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRS.R
\name{BRS_pamr.listgenes}
\alias{BRS_pamr.listgenes}
\title{List top genes from a thresholded pamr-object without forcefully printing it to R stdout}
\usage{
BRS_pamr.listgenes(fit, data, threshold, fitcv = NULL, genenames = FALSE)
}
\arguments{
\item{fit}{A fitted pamr-object by pamr.train}

\item{data}{The original data matrix provided for training}

\item{threshold}{The penalization threshold parameter}

\item{fitcv}{A pamr cross-validation object}

\item{genenames}{Inclusion of gene names taken from parameter 'data'}
}
\description{
List top genes from a thresholded pamr-object without forcefully printing it to R stdout
}
\details{
Original implementation by authors of the 'pamr'-package's function 'pamr.listgenes'; this is merely a small adjustment for user convenience.
}
