% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{BRSpred}
\alias{BRSpred}
\title{BRSpred: BCG Response Subtype Predictor for High-risk Non-Muscle Invasive Bladder Cancer}
\description{
Consensus clustering based molecular predictor for BCG responsiveness in bladder cancer. Further, transcriptomics data for training and validation are provided.
}
