% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CohortAB_counts}
\alias{CohortAB_counts}
\alias{CohortA_counts}
\alias{CohortB_counts}
\title{Gene expression matrices of raw aligned read counts in the Erasmus Medical Center bladder cancer cohort A and B}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 56912 rows and 177 columns.

An object of class \code{matrix} (inherits from \code{array}) with 35019 rows and 151 columns.
}
\usage{
CohortA_counts

CohortB_counts
}
\description{
Gene expression matrices of raw aligned read counts in the Erasmus Medical Center bladder cancer cohort A and B

CohortA and B aligned raw counts data from RNA-seq

CohortA and B aligned raw counts data from RNA-seq
}
\keyword{datasets}
