% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ErasmusMC}
\alias{ErasmusMC}
\alias{erasmus_clinical}
\alias{erasmus_immune}
\alias{CohortA_pre}
\alias{CohortA_post}
\alias{CohortB}
\title{Clinical information and gene expression matrices for the Erasmus Medical Center bladder cancer cohort A and B}
\format{
An object of class \code{data.frame} with 327 rows and 27 columns.

An object of class \code{data.frame} with 283 rows and 11 columns.

An object of class \code{data.frame} with 13922 rows and 132 columns.

An object of class \code{matrix} (inherits from \code{array}) with 16439 rows and 44 columns.

An object of class \code{data.frame} with 18733 rows and 151 columns.
}
\usage{
erasmus_clinical

erasmus_immune

CohortA_pre

CohortA_post

CohortB
}
\description{
Clinical information and gene expression matrices for the Erasmus Medical Center bladder cancer cohort A and B

erasmus_clinical: Clinical information for the Erasmus MC cohorts A (pre- and post-BCG) and B

erasmus_immune: Immune cell estimates (Sup Table S5) for the Erasmus MC cohorts A and B

CohortA_pre: ErasmusMC NMIBC Cohort A (training pre-BCG, variance stabilizing transformation)

CohortA_post: ErasmusMC NMIBC Cohort A (post-BCG, variance stabilizing transformation)

CohortB: ErasmusMC NMIBC Cohort B (validation, variance stabilizing transformation)
}
\keyword{datasets}
