% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pamr3cl}
\alias{pamr3cl}
\title{pamr 3 cluster object for prediction}
\format{
An object of class \code{pamrtrained} of length 23.
}
\usage{
pamr3cl
}
\description{
pamr 3 cluster object for prediction
}
\keyword{datasets}
