% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{BRSpred}
\alias{BRSpred}
\title{BRSpred: BCG Response Subtype Predictor for High-risk Non-Muscle Invasive Bladder Cancer}
\description{
Consensus clustering based molecular predictor was first constructed for characterizing BCG responsiveness in bladder cancer. With these initial classes, a pamr-classifier is constructed for future data. Further, transcriptomics datasets for training and validation are provided along with key clinical variables. See the main BRS vignette for further details.
}
