% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{erasmus.clinical}
\alias{erasmus.clinical}
\alias{CohortA.vst}
\alias{CohortB.vst}
\title{Clinical information for the Erasmus MC cohorts A and B}
\format{
An object of class \code{data.frame} with 328 rows and 26 columns.

An object of class \code{data.frame} with 13922 rows and 132 columns.

An object of class \code{data.frame} with 18733 rows and 151 columns.
}
\usage{
erasmus.clinical

CohortA.vst

CohortB.vst
}
\description{
Clinical information for the Erasmus MC cohorts A and B

ErasmusMC NMIBC Cohort A (training, variance stabilizing transformation)

ErasmusMC NMIBC Cohort B (validation, variance stabilizing transformation)
}
\keyword{datasets}
